fix(self-hosted): Make sure that event_id removal on cached_aggregation is safe

+++ db/migrate/20251020090137_remove_event_id_from_cached_aggregation.rb
@@ -2,8 +2,13 @@
 
 class RemoveEventIdFromCachedAggregation < ActiveRecord::Migration[8.0]
   def up
-    safety_assured do
-      remove_column :cached_aggregations, :event_id
+    # This migration has been removed to allow safe upgrade for self hosted instances
+    # It will be replaced with an other migration after a first phase of cleanup to make sure that no
+    # code relies on the event_id column anymore
+    unless Rails.env.production?
+      safety_assured do
+        remove_column :cached_aggregations, :event_id
+      end
     end
   end
 end
